# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "분리수ZIP 서비스 API",
    "description": "엘리스 AI 트랙 4기 - AI Project 10팀 TOPTEN 프로젝트 문서.",
    "version": "1.0.0"
  },
  "host": "home.handwoong.com:5001",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "http",
      "scheme": "bearer",
      "in": "header",
      "bearerFormat": "JWT"
    }
  },
  "paths": {
    "/comments/{postId}": {
      "post": {
        "tags": [
          "comment"
        ],
        "description": "댓글 생성 **로그인 필수**",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "댓글을 달 게시글의 ID",
            "schema": {
              "$ref": "#/definitions/PostId"
            }
          },
          {
            "name": "body",
            "in": "body",
            "description": "생성하고자 하는 댓글의 정보를 body에 담아 요청",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CommentCreateRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "$ref": "#/definitions/CommentCreateResponse"
            },
            "description": "생성된 댓글 정보 반환"
          }
        }
      }
    },
    "/comments/{id}": {
      "put": {
        "tags": [
          "comment"
        ],
        "description": "댓글 정보 수정 **로그인 필수**",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "수정하고자 하는 댓글의 ID",
            "schema": {
              "$ref": "#/definitions/CommentId"
            }
          },
          {
            "name": "body",
            "in": "body",
            "description": "수정하고자 하는 댓글의 정보를 body에 담아 요청",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CommentPutRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/CommentPutResponse"
            },
            "description": "수정된 댓글 정보 반환"
          }
        }
      }
    },
    "/comments/{id}/posts/{postId}": {
      "delete": {
        "tags": [
          "comment"
        ],
        "description": "댓글 삭제 **로그인 필수**",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "삭제하고자 하는 댓글의 ID",
            "schema": {
              "$ref": "#/definitions/CommentId"
            }
          },
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "삭제하고자 하는 댓글이 달려있는 게시글 ID",
            "schema": {
              "$ref": "#/definitions/PostId"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/DeleteResponse"
            },
            "description": "삭제 메시지"
          }
        }
      }
    },
...
    "/posts": {
      "get": {
        "tags": [
          "post"
        ],
        "description": "게시글 목록 조회",
        "parameters": [
          {
            "name": "queryString",
            "in": "query",
            "description": "**search** 검색어\n  **page** 첫 요청시 빈 문자열 또는 생략\n  **limit** 기본값10\n",
            "required": false,
            "schema": {
              "$ref": "#/definitions/PostGetQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/PostGetResponse"
            },
            "description": "게시글 목록을 배열형태로 반환"
          }
        }
      },
      "post": {
        "tags": [
          "post"
        ],
        "description": "게시글 생성 **로그인 필수**",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "생성하고자 하는 게시글의 정보를 body에 담아 요청",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostCreateRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "$ref": "#/definitions/PostCreateResponse"
            },
            "description": "생성된 게시글 정보 반환"
          }
        }
      }
    },
    "/posts/{id}": {
      "get": {
        "tags": [
          "post"
        ],
        "description": "개별 게시글 정보 조회",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "얻고자 하는 게시글정보의 ID",
            "schema": {
              "$ref": "#/definitions/PostId"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/PostOneGetResponse"
            },
            "description": "게시글 정보를 반환"
          }
        }
      },
      "put": {
        "tags": [
          "post"
        ],
        "description": "게시글 정보 수정 **로그인 필수**",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "수정하고자 하는 게시글의 ID",
            "schema": {
              "$ref": "#/definitions/PostId"
            }
          },
          {
            "name": "body",
            "in": "body",
            "description": "수정하고자 하는 게시글의 정보를 body에 담아 요청",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostPutRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/PostPutResponse"
            },
            "description": "수정된 게시글 정보 반환"
          }
        }
      },
      "delete": {
        "tags": [
          "post"
        ],
        "description": "게시글 삭제 **로그인 필수**",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "삭제하고자 하는 게시글의 ID",
            "schema": {
              "$ref": "#/definitions/PostId"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/DeleteResponse"
            },
            "description": "삭제 메시지"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "DeleteResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "삭제가 완료되었습니다."
        }
      }
    },
...
    "PostId": {
      "type": "string",
      "example": "posts/{postId}"
    },
    "PostGetQuery": {
      "type": "string",
      "example": "?search={}&page={ID}&limit={5}"
    },
    "PostGetResponse": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "62b469693311af4335ff10ed"
          },
          "title": {
            "type": "string",
            "example": "게시글제목"
          },
          "author": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "example": "62b265a4bec0b04d29f2c896"
              },
              "email": {
                "type": "string",
                "example": "test@test.com"
              },
              "username": {
                "type": "string",
                "example": "수정함"
              },
              "__v": {
                "type": "number",
                "example": 0
              },
              "topscore": {
                "type": "number",
                "example": 90
              }
            }
          },
          "content": {
            "type": "string",
            "example": "게시글내용"
          },
          "comments": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "62b46998c421caecd32fcd1b"
                },
                "author": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "62b265a4bec0b04d29f2c896"
                    },
                    "email": {
                      "type": "string",
                      "example": "test@test.com"
                    },
                    "username": {
                      "type": "string",
                      "example": "수정함"
                    },
                    "__v": {
                      "type": "number",
                      "example": 0
                    },
                    "topscore": {
                      "type": "number",
                      "example": 90
                    }
                  }
                },
                "content": {
                  "type": "string",
                  "example": "댓글댓글"
                },
                "createdAt": {
                  "type": "string",
                  "example": "2022-06-23T13:24:40.603Z"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "2022-06-23T13:24:40.603Z"
                },
                "__v": {
                  "type": "number",
                  "example": 0
                }
              }
            }
          },
          "createdAt": {
            "type": "string",
            "example": "2022-06-23T13:23:53.939Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2022-06-23T13:24:40.603Z"
          },
          "__v": {
            "type": "number",
            "example": 1
          }
        }
      }
    },
    "PostOneGetResponse": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "62b469693311af4335ff10ed"
        },
        "title": {
          "type": "string",
          "example": "게시글제목"
        },
        "author": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "example": "62b265a4bec0b04d29f2c896"
            },
            "email": {
              "type": "string",
              "example": "test@test.com"
            },
            "username": {
              "type": "string",
              "example": "수정함"
            },
            "__v": {
              "type": "number",
              "example": 0
            },
            "topscore": {
              "type": "number",
              "example": 90
            }
          }
        },
        "content": {
          "type": "string",
          "example": "게시글내용"
        },
        "comments": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "example": "62b46998c421caecd32fcd1b"
              },
              "author": {
                "type": "object",
                "properties": {
                  "_id": {
                    "type": "string",
                    "example": "62b265a4bec0b04d29f2c896"
                  },
                  "email": {
                    "type": "string",
                    "example": "test@test.com"
                  },
                  "username": {
                    "type": "string",
                    "example": "수정함"
                  },
                  "__v": {
                    "type": "number",
                    "example": 0
                  },
                  "topscore": {
                    "type": "number",
                    "example": 90
                  }
                }
              },
              "content": {
                "type": "string",
                "example": "댓글댓글"
              },
              "createdAt": {
                "type": "string",
                "example": "2022-06-23T13:24:40.603Z"
              },
              "updatedAt": {
                "type": "string",
                "example": "2022-06-23T13:24:40.603Z"
              },
              "__v": {
                "type": "number",
                "example": 0
              }
            }
          }
        },
        "createdAt": {
          "type": "string",
          "example": "2022-06-23T13:23:53.939Z"
        },
        "updatedAt": {
          "type": "string",
          "example": "2022-06-23T13:24:40.603Z"
        },
        "__v": {
          "type": "number",
          "example": 1
        }
      }
    },
    "PostCreateRequest": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "제목"
        },
        "content": {
          "type": "string",
          "example": "내용"
        }
      }
    },
    "PostCreateResponse": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "게시글제목"
        },
        "author": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "example": "62b265a4bec0b04d29f2c896"
            },
            "email": {
              "type": "string",
              "example": "test@test.com"
            },
            "username": {
              "type": "string",
              "example": "수정함"
            },
            "__v": {
              "type": "number",
              "example": 0
            },
            "topscore": {
              "type": "number",
              "example": 90
            }
          }
        },
        "content": {
          "type": "string",
          "example": "게시글내용"
        },
        "comment": {
          "type": "array",
          "example": [],
          "items": {}
        },
        "_id": {
          "type": "string",
          "example": "62b429a849398630ffce4e4f"
        },
        "createdAt": {
          "type": "string",
          "example": "2022-06-23T08:51:52.962Z"
        },
        "updatedAt": {
          "type": "string",
          "example": "2022-06-23T08:51:52.962Z"
        },
        "__v": {
          "type": "number",
          "example": 0
        }
      }
    },
    "PostPutRequest": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "수정할 제목"
        },
        "content": {
          "type": "string",
          "example": "수정할 내용"
        }
      }
    },
    "PostPutResponse": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "62b429a849398630ffce4e4f"
        },
        "title": {
          "type": "string",
          "example": "수정된"
        },
        "author": {
          "type": "string",
          "example": "62b265a4bec0b04d29f2c896"
        },
        "content": {
          "type": "string",
          "example": "수정된"
        },
        "comment": {
          "type": "array",
          "example": [],
          "items": {}
        },
        "createdAt": {
          "type": "string",
          "example": "2022-06-23T08:51:52.962Z"
        },
        "updatedAt": {
          "type": "string",
          "example": "2022-06-23T08:53:57.391Z"
        },
        "__v": {
          "type": "number",
          "example": 0
        }
      }
    },
    "CommentId": {
      "type": "string",
      "example": "comments/{commentId}"
    },
    "CommentCreateRequest": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string",
          "example": "댓글 내용"
        }
      }
    },
    "CommentCreateResponse": {
      "type": "object",
      "properties": {
        "author": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "example": "62b265a4bec0b04d29f2c896"
            },
            "email": {
              "type": "string",
              "example": "test@test.com"
            },
            "username": {
              "type": "string",
              "example": "수정함"
            },
            "__v": {
              "type": "number",
              "example": 0
            },
            "topscore": {
              "type": "number",
              "example": 90
            }
          }
        },
        "content": {
          "type": "string",
          "example": "댓글댓글"
        },
        "_id": {
          "type": "string",
          "example": "62b45eda265c75394cc5ba0e"
        },
        "createdAt": {
          "type": "string",
          "example": "2022-06-23T12:38:50.274Z"
        },
        "updatedAt": {
          "type": "string",
          "example": "2022-06-23T12:38:50.274Z"
        },
        "__v": {
          "type": "number",
          "example": 0
        }
      }
    },
    "CommentPutRequest": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string",
          "example": "수정할 댓글 내용"
        }
      }
    },
    "CommentPutResponse": {
      "type": "object",
      "properties": {
        "author": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "example": "62b265a4bec0b04d29f2c896"
            },
            "email": {
              "type": "string",
              "example": "test@test.com"
            },
            "username": {
              "type": "string",
              "example": "수정함"
            },
            "__v": {
              "type": "number",
              "example": 0
            },
            "topscore": {
              "type": "number",
              "example": 90
            }
          }
        },
        "content": {
          "type": "string",
          "example": "수정할 댓글 내용"
        },
        "_id": {
          "type": "string",
          "example": "62b45eda265c75394cc5ba0e"
        },
        "createdAt": {
          "type": "string",
          "example": "2022-06-23T12:38:50.274Z"
        },
        "updatedAt": {
          "type": "string",
          "example": "2022-06-23T12:38:50.274Z"
        },
        "__v": {
          "type": "number",
          "example": 0
        }
      }
    },
...
  }
}